refactor(server): drop the handleX test seams, move their tests onto ToolContext
The seven private handleX delegates on DebugMcpServer existed only so
server-coverage.test.ts could reach the handlers by name after PR 4 moved the
bodies into src/server/handlers/. They are deleted here, and the ~35 tests that
called them move to tests/core/unit/server/handlers/*.test.ts, where they call
the handler functions directly — in the same commit, so coverage never dips.

server-test-helpers.ts gains createMockToolContext(overrides?): a real
DebugMcpServer (the thing that implements ToolContext) with its session manager
and logger swapped for mocks, rather than a hand-rolled literal that would be
free to drift from the interface the handlers program against. Everything the
migrated tests stubbed on the server — fileChecker, lineReader, validateSession
— is still stubbed the same way.

One incidental fix: the migrated ProxyNotRunningError case passed one argument
to a two-argument constructor (the single error tests/unit/server-coverage.test.ts
carried in the typecheck baseline). The new files are type-clean, so the
baseline entry is dropped.

getSupportedLanguagesAsync and getLanguageMetadata stay on the class: they are
ToolContext services, not test seams.
