+++ tests/unit/server-coverage.test.ts
@@ -5,9 +5,8 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { DebugMcpServer } from '../../src/server';
-import { McpError, ErrorCode as McpErrorCode } from '@modelcontextprotocol/sdk/types.js';
+import { McpError } from '@modelcontextprotocol/sdk/types.js';
 import { SessionLifecycleState } from '@debugmcp/shared';
-import { SessionTerminatedError, ProxyNotRunningError } from '../../src/errors/debug-errors.js';
 
 describe('Server Coverage - Error Paths and Edge Cases', () => {
   let server: DebugMcpServer;
@@ -612,40 +611,6 @@ describe('Server Coverage - Error Paths and Edge Cases', () => {
       await expect(server.stepInto('test-session')).resolves.toEqual({ success: true, state: 'paused' });
       await expect(server.stepOut('test-session')).resolves.toEqual({ success: true, state: 'paused' });
     });
-
-    it('handleListDebugSessions maps active sessions', async () => {
-      const now = new Date();
-      mockSessionManager.getAllSessions.mockReturnValue([{
-        id: 'session-1',
-        name: 'Test Session',
-        language: 'python',
-        state: 'active',
-        createdAt: now,
-        updatedAt: now
-      }]);
-
-      const result = await (server as any).handleListDebugSessions();
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.count).toBe(1);
-      expect(payload.sessions[0]).toMatchObject({
-        id: 'session-1',
-        name: 'Test Session',
-        language: 'python'
-      });
-    });
-
-    it('handlePause returns success:false when session validation fails', async () => {
-      mockSessionManager.pause = vi.fn().mockRejectedValue(new Error('some pause error'));
-      mockSessionManager.getSession.mockReturnValue(null);
-
-      const result = await (server as any).handlePause({ sessionId: 'test-session' });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('Session not found: test-session');
-    });
   });
 
   describe('Get Session Name Error Handling', () => {
@@ -810,617 +775,6 @@ describe('Server Coverage - Error Paths and Edge Cases', () => {
           condition: undefined, suspendPolicy: undefined, logMessage: undefined
         })
       );
-    });
-  });
-
-  describe('handleListThreads', () => {
-    it('should return threads on success', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.listThreads = vi.fn().mockResolvedValue([
-        { id: 1, name: 'main' },
-        { id: 2, name: 'worker-1' },
-      ]);
-
-      const result = await (server as any).handleListThreads({ sessionId: 'test-session' });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.threads).toHaveLength(2);
-      expect(payload.threads[0]).toEqual({ id: 1, name: 'main' });
-    });
-
-    it('converts typed session errors (SessionTerminatedError etc.) into success:false results', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      const { SessionTerminatedError } = await import('../../src/errors/debug-errors');
-      mockSessionManager.listThreads = vi.fn().mockRejectedValue(new SessionTerminatedError('test-session'));
-
-      const result = await (server as any).handleListThreads({ sessionId: 'test-session' });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('Session is terminated: test-session');
-    });
-
-    it('should throw McpError for unknown errors', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.listThreads = vi.fn().mockRejectedValue(new Error('unexpected'));
-
-      await expect((server as any).handleListThreads({ sessionId: 'test-session' }))
-        .rejects.toThrow('Failed to list threads: unexpected');
-    });
-  });
-
-  describe('handlePause with threadId', () => {
-    it('should pass threadId to session manager pause', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.pause = vi.fn().mockResolvedValue({ success: true, state: 'paused' });
-
-      const result = await (server as any).handlePause({ sessionId: 'test-session', threadId: 7 });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(mockSessionManager.pause).toHaveBeenCalledWith('test-session', 7);
-    });
-
-    it('should pass undefined threadId when not provided', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.pause = vi.fn().mockResolvedValue({ success: true, state: 'paused' });
-
-      await (server as any).handlePause({ sessionId: 'test-session' });
-
-      expect(mockSessionManager.pause).toHaveBeenCalledWith('test-session', undefined);
-    });
-  });
-
-  describe('Evaluate Expression Edge Cases', () => {
-    it('should handle expression evaluation in terminated session', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.TERMINATED
-      });
-
-      const result = await (server as any).handleEvaluateExpression({
-        sessionId: 'test-session',
-        expression: 'x + 1'
-      });
-
-      // The method returns a success response with the error in the content
-      expect(result.content[0].text).toContain('Session is terminated');
-    });
-  });
-
-  describe('handlePause', () => {
-    it('returns result on success', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.pause = vi.fn().mockResolvedValue({ success: true, state: 'paused' });
-
-      const result = await (server as any).handlePause({ sessionId: 'test-session' });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.state).toBe('paused');
-    });
-
-    it('converts SessionTerminatedError into a success:false result', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.pause = vi.fn().mockRejectedValue(new SessionTerminatedError('test-session'));
-
-      const result = await (server as any).handlePause({ sessionId: 'test-session' });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('Session is terminated: test-session');
-    });
-
-    it('converts ProxyNotRunningError into a success:false result', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.pause = vi.fn().mockRejectedValue(new ProxyNotRunningError('test-session'));
-
-      const result = await (server as any).handlePause({ sessionId: 'test-session' });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('test-session');
-    });
-
-    it('wraps generic errors as McpError', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.pause = vi.fn().mockRejectedValue(new Error('unexpected'));
-
-      await expect((server as any).handlePause({ sessionId: 'test-session' }))
-        .rejects.toThrow('Failed to pause execution');
-    });
-  });
-
-  describe('handleGetSourceContext', () => {
-    beforeEach(() => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-    });
-
-    it('returns source context on success', async () => {
-      (server as any).fileChecker = {
-        checkExists: vi.fn().mockResolvedValue({ exists: true, effectivePath: '/app/main.py' })
-      };
-      (server as any).lineReader = {
-        getLineContext: vi.fn().mockResolvedValue({
-          lineContent: 'print("hello")',
-          surrounding: [
-            { line: 9, content: 'def main():' },
-            { line: 10, content: '    print("hello")' },
-            { line: 11, content: '' }
-          ]
-        })
-      };
-
-      const result = await (server as any).handleGetSourceContext({
-        sessionId: 'test-session',
-        file: '/app/main.py',
-        line: 10
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.file).toBe('/app/main.py');
-      expect(payload.line).toBe(10);
-      expect(payload.lineContent).toBe('print("hello")');
-      expect(payload.surrounding).toHaveLength(3);
-    });
-
-    it('throws McpError when file not found', async () => {
-      (server as any).fileChecker = {
-        checkExists: vi.fn().mockResolvedValue({
-          exists: false,
-          effectivePath: '/missing.py',
-          errorMessage: 'ENOENT'
-        })
-      };
-
-      await expect((server as any).handleGetSourceContext({
-        sessionId: 'test-session',
-        file: '/missing.py',
-        line: 1
-      })).rejects.toThrow('Source file');
-    });
-
-    it('returns error JSON when file is unreadable', async () => {
-      (server as any).fileChecker = {
-        checkExists: vi.fn().mockResolvedValue({ exists: true, effectivePath: '/app/binary.dat' })
-      };
-      (server as any).lineReader = {
-        getLineContext: vi.fn().mockResolvedValue(null)
-      };
-
-      const result = await (server as any).handleGetSourceContext({
-        sessionId: 'test-session',
-        file: '/app/binary.dat',
-        line: 1
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('Could not read source context');
-    });
-
-    it('uses default context lines when not specified', async () => {
-      (server as any).fileChecker = {
-        checkExists: vi.fn().mockResolvedValue({ exists: true, effectivePath: '/app/test.py' })
-      };
-      const getLineContext = vi.fn().mockResolvedValue({
-        lineContent: 'x = 1',
-        surrounding: []
-      });
-      (server as any).lineReader = { getLineContext };
-
-      await (server as any).handleGetSourceContext({
-        sessionId: 'test-session',
-        file: '/app/test.py',
-        line: 5
-      });
-
-      expect(getLineContext).toHaveBeenCalledWith('/app/test.py', 5, { contextLines: 5 });
-    });
-  });
-
-  describe('handleGetLocalVariables', () => {
-    beforeEach(() => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-      mockSessionManager.getLocalVariables = vi.fn();
-    });
-
-    it('returns variables with frame and scope info', async () => {
-      mockSessionManager.getLocalVariables.mockResolvedValue({
-        variables: [{ name: 'x', value: '42' }],
-        frame: { name: 'main', file: 'test.py', line: 10 },
-        scopeName: 'Locals'
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.count).toBe(1);
-      expect(payload.variables[0].name).toBe('x');
-      expect(payload.frame.name).toBe('main');
-      expect(payload.scopeName).toBe('Locals');
-    });
-
-    it('shows "not paused" message when no frame available', async () => {
-      mockSessionManager.getLocalVariables.mockResolvedValue({
-        variables: [],
-        frame: null,
-        scopeName: null
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.count).toBe(0);
-      expect(payload.message).toContain('No stack frames available');
-    });
-
-    it('shows "no local scope" message when frame exists but no scope', async () => {
-      mockSessionManager.getLocalVariables.mockResolvedValue({
-        variables: [],
-        frame: { name: 'main', file: 'test.py', line: 10 },
-        scopeName: null
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.message).toContain('No local scope found');
-    });
-
-    it('shows "scope is empty" message when scope exists but has no variables', async () => {
-      mockSessionManager.getLocalVariables.mockResolvedValue({
-        variables: [],
-        frame: { name: 'main', file: 'test.py', line: 10 },
-        scopeName: 'Locals'
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.message).toContain('The Locals scope is empty');
-    });
-
-    it('surfaces a warning when the scope name carries an adapter warning (optimized Go binary)', async () => {
-      mockSessionManager.getLocalVariables.mockResolvedValue({
-        variables: [],
-        frame: { name: 'main.main', file: 'main.go', line: 10 },
-        scopeName: 'Locals (warning: optimized function)'
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.scopeName).toBe('Locals (warning: optimized function)');
-      expect(payload.warning).toContain('optimizations');
-      expect(payload.warning).toContain('-gcflags');
-    });
-
-    it('does not add a warning for a plain scope name', async () => {
-      mockSessionManager.getLocalVariables.mockResolvedValue({
-        variables: [{ name: 'x', value: '42' }],
-        frame: { name: 'main', file: 'test.py', line: 10 },
-        scopeName: 'Locals'
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.warning).toBeUndefined();
-    });
-
-    it('returns graceful JSON for McpError with "not paused"', async () => {
-      (server as any).validateSession = vi.fn().mockImplementation(() => {
-        throw new McpError(McpErrorCode.InvalidRequest, 'Session is not paused');
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('not paused');
-      expect(payload.message).toContain('Cannot get local variables');
-    });
-
-    it('explains a terminated session as a normal end state (program finished)', async () => {
-      (server as any).validateSession = vi.fn().mockImplementation(() => {
-        throw new McpError(McpErrorCode.InvalidRequest, 'Session is terminated: test-session');
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(false);
-      expect(payload.message).toContain('program has terminated');
-      expect(payload.message).toContain('restart_debugging');
-    });
-
-    it('wraps generic errors as McpError', async () => {
-      mockSessionManager.getLocalVariables.mockRejectedValue(new Error('unexpected'));
-
-      await expect((server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      })).rejects.toThrow('Failed to get local variables');
-    });
-  });
-
-  describe('handleListSupportedLanguages', () => {
-    it('returns installed languages and adapter metadata', async () => {
-      const result = await (server as any).handleListSupportedLanguages();
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.installed).toEqual(['python', 'mock']);
-      expect(payload.available).toHaveLength(2);
-      expect(payload.available[0].language).toBe('python');
-      expect(payload.available[0].package).toBe('@debugmcp/adapter-python');
-      expect(payload.count).toBe(2);
-    });
-
-    it('falls back to installed list when listAvailableAdapters fails', async () => {
-      mockSessionManager.adapterRegistry.listAvailableAdapters.mockRejectedValue(
-        new Error('metadata unavailable')
-      );
-
-      const result = await (server as any).handleListSupportedLanguages();
-      const payload = JSON.parse(result.content[0].text);
-
-      expect(payload.success).toBe(true);
-      expect(payload.installed).toEqual(['python', 'mock']);
-      // available falls back to simple format derived from installed
-      expect(payload.available).toHaveLength(2);
-      expect(payload.available[0].installed).toBe(true);
-    });
-  });
-
-  describe('handleGetSourceContext', () => {
-    it('returns binary/inaccessible message when lineReader returns null', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-
-      (server as any).fileChecker = {
-        checkExists: vi.fn().mockResolvedValue({
-          exists: true,
-          effectivePath: '/path/to/binary.bin'
-        })
-      };
-
-      (server as any).lineReader = {
-        getLineContext: vi.fn().mockResolvedValue(null)
-      };
-
-      const result = await (server as any).handleGetSourceContext({
-        sessionId: 'test-session',
-        file: '/path/to/binary.bin',
-        line: 1
-      });
-
-      const payload = JSON.parse(result.content[0].text);
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('binary or inaccessible');
-    });
-
-    it('returns source context when lineReader returns content', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE,
-        name: 'my-session'
-      });
-
-      (server as any).fileChecker = {
-        checkExists: vi.fn().mockResolvedValue({
-          exists: true,
-          effectivePath: '/path/to/script.py'
-        })
-      };
-
-      (server as any).lineReader = {
-        getLineContext: vi.fn().mockResolvedValue({
-          lineContent: 'x = 42',
-          surrounding: ['', 'x = 42', '']
-        })
-      };
-
-      const result = await (server as any).handleGetSourceContext({
-        sessionId: 'test-session',
-        file: '/path/to/script.py',
-        line: 5,
-        linesContext: 3
-      });
-
-      const payload = JSON.parse(result.content[0].text);
-      expect(payload.success).toBe(true);
-      expect(payload.lineContent).toBe('x = 42');
-      expect(payload.contextLines).toBe(3);
-    });
-
-    it('throws when file does not exist', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE
-      });
-
-      (server as any).fileChecker = {
-        checkExists: vi.fn().mockResolvedValue({
-          exists: false,
-          effectivePath: '/nope.py',
-          errorMessage: 'not found'
-        })
-      };
-
-      await expect((server as any).handleGetSourceContext({
-        sessionId: 'test-session',
-        file: '/nope.py',
-        line: 1
-      })).rejects.toThrow();
-    });
-  });
-
-  describe('handleGetLocalVariables', () => {
-    it('returns "no stack frames" message when frame is null', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE,
-        name: 'my-session'
-      });
-
-      mockSessionManager.getLocalVariables = vi.fn().mockResolvedValue({
-        variables: [],
-        frame: null,
-        scopeName: null
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-
-      const payload = JSON.parse(result.content[0].text);
-      expect(payload.success).toBe(true);
-      expect(payload.variables).toEqual([]);
-      expect(payload.message).toContain('No stack frames available');
-    });
-
-    it('returns "no local scope" message when scopeName is null', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE,
-        name: 'my-session'
-      });
-
-      mockSessionManager.getLocalVariables = vi.fn().mockResolvedValue({
-        variables: [],
-        frame: { name: 'main', file: 'test.py', line: 1 },
-        scopeName: null
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-
-      const payload = JSON.parse(result.content[0].text);
-      expect(payload.message).toContain('No local scope found');
-    });
-
-    it('returns "scope is empty" message when scope exists but has no variables', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE,
-        name: 'my-session'
-      });
-
-      mockSessionManager.getLocalVariables = vi.fn().mockResolvedValue({
-        variables: [],
-        frame: { name: 'main', file: 'test.py', line: 1 },
-        scopeName: 'Locals'
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-
-      const payload = JSON.parse(result.content[0].text);
-      expect(payload.message).toContain('Locals scope is empty');
-    });
-
-    it('returns variables with frame and scope info', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE,
-        name: 'my-session'
-      });
-
-      mockSessionManager.getLocalVariables = vi.fn().mockResolvedValue({
-        variables: [{ name: 'x', value: '42', type: 'int' }],
-        frame: { name: 'main', file: 'test.py', line: 10 },
-        scopeName: 'Locals'
-      });
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-
-      const payload = JSON.parse(result.content[0].text);
-      expect(payload.success).toBe(true);
-      expect(payload.variables).toHaveLength(1);
-      expect(payload.frame.name).toBe('main');
-      expect(payload.scopeName).toBe('Locals');
-      expect(payload.message).toBeUndefined();
-    });
-
-    it('returns graceful error for terminated session', async () => {
-      mockSessionManager.getSession.mockReturnValue({
-        id: 'test-session',
-        sessionLifecycle: SessionLifecycleState.ACTIVE,
-        name: 'my-session'
-      });
-
-      const { McpError: RealMcpError, ErrorCode } = await import('@modelcontextprotocol/sdk/types.js');
-      mockSessionManager.getLocalVariables = vi.fn().mockRejectedValue(
-        new RealMcpError(ErrorCode.InvalidRequest, 'Session is terminated: test-session')
-      );
-
-      const result = await (server as any).handleGetLocalVariables({
-        sessionId: 'test-session'
-      });
-
-      const payload = JSON.parse(result.content[0].text);
-      expect(payload.success).toBe(false);
-      expect(payload.error).toContain('terminated');
     });
   });
 });
